+++ logs/winston-error.log
@@ -226,3 +226,17 @@
 {"message":"Not Found","level":"error","service":"rest-service","timestamp":"2021-06-23 10:32:29"}
 {"message":"Not Found","level":"error","service":"rest-service","timestamp":"2021-06-23 10:32:29"}
 {"message":"Not Found","level":"error","service":"rest-service","timestamp":"2021-06-23 10:32:31"}
+{"message":"Not Found","level":"error","service":"rest-service","timestamp":"2021-06-23 12:49:47"}
+{"message":"Not Found","level":"error","service":"rest-service","timestamp":"2021-06-23 12:49:48"}
+{"message":"Not Found","level":"error","service":"rest-service","timestamp":"2021-06-23 12:49:48"}
+{"message":"Not Found","level":"error","service":"rest-service","timestamp":"2021-06-23 12:49:48"}
+{"message":"Not Found","level":"error","service":"rest-service","timestamp":"2021-06-23 12:49:48"}
+{"message":"Not Found","level":"error","service":"rest-service","timestamp":"2021-06-23 12:49:48"}
+{"message":"Not Found","level":"error","service":"rest-service","timestamp":"2021-06-23 12:49:48"}
+{"message":"Not Found","level":"error","service":"rest-service","timestamp":"2021-06-24 11:34:32"}
+{"message":"Not Found","level":"error","service":"rest-service","timestamp":"2021-06-24 11:34:33"}
+{"message":"Not Found","level":"error","service":"rest-service","timestamp":"2021-06-24 11:34:33"}
+{"message":"Not Found","level":"error","service":"rest-service","timestamp":"2021-06-24 11:34:33"}
+{"message":"Not Found","level":"error","service":"rest-service","timestamp":"2021-06-24 11:34:33"}
+{"message":"Not Found","level":"error","service":"rest-service","timestamp":"2021-06-24 11:34:33"}
+{"message":"Not Found","level":"error","service":"rest-service","timestamp":"2021-06-24 11:34:33"}
