+++ ormconfig.ts
@@ -1,10 +1,10 @@
 import { ConnectionOptions } from 'typeorm';
-import { POSTGRES_HOST, POSTGRES_USER, POSTGRES_PASSWORD, POSTGRES_DB } from './src/common/config';
+import { POSTGRES_HOST, POSTGRES_USER, POSTGRES_PORT, POSTGRES_PASSWORD, POSTGRES_DB } from './src/common/config';
 
 const ormconfig: ConnectionOptions = {
     type: 'postgres',
     host: POSTGRES_HOST,
-    port: 5432,
+    port: +POSTGRES_PORT,
     username: POSTGRES_USER,
     password: POSTGRES_PASSWORD,
     database: POSTGRES_DB,
